Execute function inside tracee.
Implements ExecuteInProcess to executes a function from a lib in the
tracee. Also moves the functions for attaching and injecting a lib
into the publicly visable include directory.

Bug: http://b/184954756
Test: Unit test in PR.
