+++ src/UserSpaceInstrumentation/ExecuteMachineCode.h
@@ -1,0 +1,27 @@
+// Copyright (c) 2021 The Orbit Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#ifndef USER_SPACE_INSTRUMENTATION_EXECUTE_MACHINE_CODE_H_
+#define USER_SPACE_INSTRUMENTATION_EXECUTE_MACHINE_CODE_H_
+
+#include <sys/types.h>
+
+#include <cstdint>
+
+#include "MachineCode.h"
+#include "OrbitBase/Result.h"
+
+namespace orbit_user_space_instrumentation {
+
+// Copies `code` to `address_code` in the tracee and executes it. The memory at `address_code` needs
+// to be allocated using AllocateInTracee. The code segment has to end with an `int3`. Takes care of
+// backup and restore of register state in the tracee and also deallocates the memory at
+// `address_code` afterwards. The return value is the content of rax after the execution finished.
+[[nodiscard]] ErrorMessageOr<uint64_t> ExecuteMachineCode(pid_t pid, uint64_t address_code,
+                                                          uint64_t memory_size,
+                                                          const MachineCode& code);
+
+}  // namespace orbit_user_space_instrumentation
+
+#endif  // USER_SPACE_INSTRUMENTATION_EXECUTE_MACHINE_CODE_H_
