+++ src/UserSpaceInstrumentation/TrampolineTest.cpp
@@ -12,11 +12,11 @@
 #include <vector>
 
 #include "AllocateInTracee.h"
-#include "Attach.h"
 #include "ElfUtils/LinuxMap.h"
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/ReadFileToString.h"
 #include "Trampoline.h"
+#include "UserSpaceInstrumentation/Attach.h"
 
 namespace orbit_user_space_instrumentation {
