+++ src/UserSpaceInstrumentation/UserSpaceInstrumentationTestLib.cpp
@@ -5,3 +5,7 @@
 #include "UserSpaceInstrumentationTestLib.h"
 
 int TrivialFunction() { return 42; }
+
+uint64_t TrivialSum(uint64_t p0, uint64_t p1, uint64_t p2, uint64_t p3, uint64_t p4, uint64_t p5) {
+  return p0 + p1 + p2 + p3 + p4 + p5;
+}
